+++ spec/setup.js
@@ -2,3 +2,4 @@ var chai = require( 'chai' );
 chai.use( require( 'chai-as-promised' ) );
 global.should = chai.should();
 global.sinon = require( 'sinon' );
+require( 'sinon-as-promised' );